requirements.txt
flask
requests
requests_mock
testing.redis
bs4
